# Add more optimised find methods for BioSequences